Move offer modal to separate file
